Move Home page to HomeLayout component
The homepage in index.js calls on a HomeLayout component that
controls the layout of the homepage. Having this Component rather
than putting the home page components in the index.js file
allows for the styling file to be labeled in a way that is not
confusing, rather than index.module.css, there is a HomeLayout.modules.
css