feat: Mobile Nav Menu
Add Mobile Nav Menu. Use Font Awesome for menu icon, which toggles
the display of the topic sections, that drop down below the top
menu section. When clicking on a navlink on the menu then menu
will hidden and when clicking on the section link in the menu,
the menu will hide and load the new page.