Feat: Scroll To Topics on Start Learning Button
When the user clicks the start learning button in the hero of the
home page the page will automatically scroll to the topics section.
Using a useRef hook in the topics component to reference the topics
div element. UseEffect Hook which dependencies are a scrollToTopics
flagand the ref, has a function that scrolls to the ref element.

When the component has the scrollToTopics flag updated to true the
scroll occurs. The flag is a prop that is from the a state in the
Home layout component that is changed on the click on the start
learning button that is in the hero component.